added faux-article content
Downloaded js pre-requisites so we host them internally for now, in the
event an external site goes down or we test internally (CORS exceptions
can be avoided).

Brought in (new) cleaned-up jumbotron video [5mb at ~25sec on loop].
Set navbar to fixed, at top.
Brought in fake article content.
Used jQuery load on document.ready to bring in articles, for future
enhancement we can make this a fade-in scroll effect when enough content
supports it.